move added UT to mor-delete test
